chore: move localcatalog package from pkg to internal
