+++ cli/internal/localcatalog/catalog.go
