+++ cli/internal/localcatalog/catalog_test.go
