+++ cli/internal/localcatalog/testdata/events_batch_1.yaml
