+++ cli/internal/localcatalog/testdata/events_batch_2.yaml
