+++ cli/internal/localcatalog/testdata/properties_batch_1.yaml
