+++ cli/internal/localcatalog/testdata/properties_batch_2.yaml
