+++ cli/internal/validate/custom_types_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	catalog "github.com/rudderlabs/rudder-iac/cli/pkg/localcatalog"
+	catalog "github.com/rudderlabs/rudder-iac/cli/internal/localcatalog"
 	"github.com/stretchr/testify/assert"
 )
