+++ cli/internal/validate/validate.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/rudderlabs/rudder-iac/cli/internal/localcatalog"
 	"github.com/rudderlabs/rudder-iac/cli/internal/logger"
-	"github.com/rudderlabs/rudder-iac/cli/pkg/localcatalog"
 )
 
 var log = logger.New("validate")
